# docs/swagger.yaml
...
openapi: 3.0.0
...
info:
  title: Swagger Tutorial
  description: '**Tutorial** de `Swagger` para *documentación* de API escrita con JavaScript, para más información visite la *documentación* de [**swagger.io**](https://swagger.io/)'
  termsOfService: https://swagger.io/docs/specification/about/
  contact:
    name: emenpe
    url: https://swagger.io/docs/specification/about/
    email: emenpe@emenpe.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0

#################### BRINDAR MAS INFORMACIÓN AL USUARIO CON DOCUMENTACION DE TERCEROS ####################
...
servers:
- url: http://localhost:8080/

#################### TARGETAS QUE CONTIENEN LAS RUTAS QUE SERAN DEFINIDAS ####################
...
paths:
...
  /api/users:
    get:
      tags:
      - user
      summary: Obtener Usuarios
      description: Permite obtener todos los usuarios
      operationId: getUsers
      responses:
        "200":
          description: Usuarios obtenidos
          content:
            aplication/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        "401":
          description: No autorizado
          content:
            aplication/json:
              examples:
                token_nulo:
                  $ref: '#/components/examples/token_nulo'
                token_Expirado:
                  $ref: '#/components/examples/token_Expirado'
                token_Malformado:
                  $ref: '#/components/examples/token_Malformado'
                token_Invalido:
                  $ref: '#/components/examples/token_Invalido'
        "500":
          description: Error del lado del servidor
          content:
            application/json:
              example:
                msg: Error al obtener los usuarios
      security:
      - bearerAuth: []
...
components:
...
  schemas:
    User:
      type: object
      properties:
        name:
          type: string
          description: Descripción de este elemento del objeto
        email:
          type: string
          description: Descripción de este elemento del objeto
        password:
          type: string
          description: Descripción de este elemento del objeto
      example:
        name: fulano
        email: fulano@fulano.com
        password: fulano12345

  #################### DEFINICION DE EJEMPLOS DE RESPUESTAS QUE PUDE MANDAR EL SERVIDOR   #################### 
  examples:
    token_nulo:
      summary: Token_nulo
      description: Este error ocurre cuando el token no es enviado
      value:
        msg: Token nulo
    token_Expirado:
      summary: Token_expirado
      description: Este error ocurre cuando el token proporcionado ha expirado
      value:
        msg: Token Expirado
    token_Malformado:
      summary: Token_Malformado
      description: Este error ocurre cuando el token proporcionado tiene un error en su estructura
      value:
        msg: Token malformado
    token_Invalido:
      summary: Token_Invalido
      description: Este error ocurre cuando el token proporcionado es inválido por algunas razones en la validación
      value:
        msg: Token invalido
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  #################### DEFINICION DE LINKS DONDE EL VALOR DEL RESPONSE PUEDE USARSE PARA REALIZAR ALGUNAS OTRAS PETICIONES EN LAS RUTAS  #################### 
...